Closes #3 new Binance order placement

+++ src/main/java/cz/amuradon/tvbot/WebhookData.java
@@ -1,10 +1,8 @@
 package cz.amuradon.tvbot;
 
-import java.math.BigDecimal;
-
 public record WebhookData(String symbol,
 		String side,
-		BigDecimal quantity,
+		double quantity,
 		boolean reduceOnly,
 		String newClientOrderId,
 		String userUuid) {
